refactor: improve Table component
